Encode exp as integer in app jwt.
Fixes 401 errors in current github api due to use of floating point,
rather than integer value in app jwt.

Github, y u change api behavior? ಠ_ಠ

+++ git_credential_github_app_auth/identity.py
@@ -81,8 +81,8 @@ class AppIdentity:
         """
         issue_time = time.time() - 1
         payload = dict(
-          iat= issue_time,
-          exp= issue_time + (10 * 60),
+          iat= int(issue_time),
+          exp= int(issue_time + (10 * 60)),
           iss= self.app_id
         )
